+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.1.7
+
+- Added `CSSFunction` (extends `CSSValues`):
+  - New `CSSValues`: `CSSMax` and `CSSMin`.
+  - `CSSCalc` now extends `CSSFunction`.
+
 ## 2.1.6
 
 - Fix `createTableCells`.